fix chatbox

+++ src/main/java/com/watch/shopwatchonline/Controller/Site/wishlistController.java
@@ -9,6 +9,7 @@ import java.util.stream.Collectors;
 import java.util.stream.IntStream;
 
 import javax.servlet.http.HttpServletRequest;
+import javax.servlet.http.HttpServletResponse;
 
 import org.springframework.beans.BeanUtils;
 import org.springframework.beans.factory.annotation.Autowired;
@@ -16,6 +17,7 @@ import org.springframework.data.domain.Page;
 import org.springframework.data.domain.PageRequest;
 import org.springframework.data.domain.Pageable;
 import org.springframework.data.domain.Sort;
+import org.springframework.http.ResponseEntity;
 import org.springframework.stereotype.Controller;
 import org.springframework.ui.ModelMap;
 import org.springframework.util.StringUtils;
@@ -23,8 +25,10 @@ import org.springframework.web.bind.annotation.CrossOrigin;
 import org.springframework.web.bind.annotation.GetMapping;
 import org.springframework.web.bind.annotation.PathVariable;
 import org.springframework.web.bind.annotation.PostMapping;
+import org.springframework.web.bind.annotation.RequestBody;
 import org.springframework.web.bind.annotation.RequestMapping;
 import org.springframework.web.bind.annotation.RequestParam;
+import org.springframework.web.bind.annotation.ResponseBody;
 import org.springframework.web.servlet.ModelAndView;
 
 import com.watch.shopwatchonline.Model.Product;
@@ -42,18 +46,20 @@ import com.watch.shopwatchonline.security.jwt.JwtUtils;
 public class wishlistController {
     @Autowired
     private WishlistRepository wishlistRepository;
-    @Autowired private ProductService ProductService;
-    @Autowired private JwtUtils jwtUtils;
-    @Autowired private UserRepository userRepository;
+    @Autowired
+    private ProductService ProductService;
+    @Autowired
+    private JwtUtils jwtUtils;
+    @Autowired
+    private UserRepository userRepository;
 
-    @GetMapping("") public ModelAndView index(ModelMap model,HttpServletRequest request,
-        @RequestParam("page") Optional < Integer > page,
-        @RequestParam("size") Optional < Integer > size,
-        @RequestParam("sort") Optional < Integer > sort) {
+    @GetMapping("")
+    public ModelAndView index(ModelMap model, HttpServletRequest request,
+            @RequestParam("page") Optional<Integer> page,
+            @RequestParam("size") Optional<Integer> size,
+            @RequestParam("sort") Optional<Integer> sort) {
 
-          
-       
-        Page < Product > resultPage = null;
+        Page<Product> resultPage = null;
 
         int curPage = page.orElse(1);
         int pageSize = size.orElse(6);
@@ -80,15 +86,13 @@ public class wishlistController {
 
             Pageable pageable = PageRequest.of(curPage - 1, pageSize, sortable);
 
-       
-
-            resultPage = ProductService.findByUserName(getUser(request),pageable);
+            resultPage = ProductService.findByUserName(getUser(request), pageable);
             model.addAttribute("sort", abx);
 
         } else {
             Pageable pageable = PageRequest.of(curPage - 1, pageSize);
 
-            resultPage = ProductService.findByUserName(getUser(request),pageable);
+            resultPage = ProductService.findByUserName(getUser(request), pageable);
 
         }
 
@@ -108,7 +112,7 @@ public class wishlistController {
                 }
             }
 
-            List < Integer > pageNums = IntStream.rangeClosed(start, end).boxed().collect(Collectors.toList());
+            List<Integer> pageNums = IntStream.rangeClosed(start, end).boxed().collect(Collectors.toList());
             model.addAttribute("pageNums", pageNums);
 
         }
@@ -117,8 +121,6 @@ public class wishlistController {
         return new ModelAndView("web-site/wishlist", model);
     }
 
-
-
     @GetMapping("delete/{Id}")
     public String delete(ModelMap map, @PathVariable("Id") Integer id) {
         wishlistRepository.deleteById(id);
@@ -126,20 +128,22 @@ public class wishlistController {
         return "redirect:/api/admin/raiting/list-waiting";
     }
 
-    @GetMapping("like/{Id}")
-    public String store(ModelMap model, @PathVariable("Id") Integer id, HttpServletRequest request) {
-        int like;
-        try {
-            wishlist wl = new wishlist();
-            Optional < Product > opt = ProductService.findById(id);
-            Optional < User > user = userRepository.findByUsername(getUser(request));
+    
+    @GetMapping("/like")
+    public  ResponseEntity<?> saveOrUpdateCompany(
+        @RequestBody @RequestParam(name = "username", required = false) String username,
+        @RequestBody @RequestParam(name = "id") Integer id) {
+  
+      try {
+        wishlist wl = new wishlist();
+        Optional<Product> opt = ProductService.findById(id);
+        Optional<User> user = userRepository.findByUsername(username);
+        if (user.isPresent()) {
             if (!wishlistRepository.check(opt.get().getId(), user.get().getId()).isEmpty()) {
 
-                wishlistRepository.deleteByProAndUser(opt.get().getId(), user.get().getId());
+                 wishlistRepository.deleteByProAndUser(opt.get().getId(), user.get().getId());
 
-                like = 0;
-                model.addAttribute("like", like);
-
+                return  ResponseEntity.ok().body(0);
             } else {
 
                 wl.setProduct(opt.get());
@@ -148,26 +152,25 @@ public class wishlistController {
                 wl.setCreateAt(new Date());
                 wishlistRepository.save(wl);
 
-                like = 1;
-                model.addAttribute("like", like);
+              
+                return  ResponseEntity.ok().body(1);
             }
-
-
-
-        } catch (Exception e) {
-            System.out.println("+++++++++++++++++++++++++");
-            System.out.println(e);
+        } else {
+            return  ResponseEntity.ok().body("null");
         }
+      } catch (Exception e) {
+        return  ResponseEntity.ok().body(0);
+      }
 
-
-
-        return "redirect:/api/site/product/detail/"+id ;
     }
 
-    public String getUser( HttpServletRequest request){
+    public String getUser(HttpServletRequest request) {
         String token = jwtUtils.getJwtFromCookies(request);
         String username = jwtUtils.getUserNameFromJwtToken(token);
-       
+
         return username;
     }
+
+
+    //@ResponseBody get đc path 
 }
